Added a property to quote grid

Related work items: #191377

+++ services/Commerce/DigitalCommercePlatform.UIServices.Commerce/Models/Quote/Quote/Internal/ItemModelBase.cs
@@ -9,7 +9,7 @@ namespace DigitalCommercePlatform.UIServices.Commerce.Models.Quote.Quote.Interna
         public string Id { get; set; }
         public string Parent { get; set; }
         public List<ProductModel> Product { get; set; }
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public string ContractNumber { get; set; }
         public string ContractType { get; set; }
         public string License { get; set; }
@@ -20,6 +20,5 @@ namespace DigitalCommercePlatform.UIServices.Commerce.Models.Quote.Quote.Interna
         public decimal UnitListPrice { get; set; }
         public decimal ExtendedListPrice { get; set; }
         public EndUserModel EndUser { get; set; }
-
     }
 }
